Infrastructure for interaction (impedance) control
Implementing the InteractionController plugin, which includes
the basic infrastructure for cartesian impedance control.
The cartesian impedance control can be turned on/off with
desired stiffness and damping values by publishing a rostopic.
This can be used for doing constrained zero-G.

Reviewboard adventures! http://reviewboard.rethink.cxm/r/5746/

Drafts of documents about this can be found in:
https://drive.google.com/open?id=1wyPN3eTakBd_-v_cupR7M7RE8PzYidr_oDJEKKdmd9U
https://drive.google.com/open?id=1nkhve3n8zCoqIWFXFj8dLS3W6DI5OZFwiiuZPwxgM0Y

This commit will be useful after the robot_config changes
are merged in.