Fix false collisions for new express accelerations
The new acceleration limits on express were causing false collisions on
existing tasks. There is logic in the detector, which is a holdover from
Baxter, that only updates the expected (feedforward) torque when the robot
was moving through a trajectory, otherwise the expected torque was set to 0.
This is OK for most situations, but on the express settings there were
situations where the state determining that the robot was in a trajectory was
not being updated before the robot started moving and the detector was
comparing the measured torque to 0.0 rather than the feedforward torque,
causing a false collision. I've removed that logic; now we always compare the
the measured torque to the feedforward torque for collision detection purposes.

Resolves #17174
review: http://reviewboard.rethink.cxm/r/6848/