Communicate the commanded force from the interaction controller
In many tasks, it would be useful to see the value of
commanded force (wrench) that is generated by the
interaction controller. This commit sends this infromation
through a rostopic. The engine can then listen to these
values and commanded force can be used in the
behavior tree to do conditionals.

review: http://reviewboard.rethink.cxm/r/7000/

resolves #17461
resolves #17487