Impedance modulation using quadratic programming
This code implements impedance modulation, which is an
algorithm that automatically solves for the maximum stiffness
given the pose of the robot. This algorithm uses a quadratic
program to solve for the optimal stiffness values.

We are using an open source quadratic program solver with
an MIT license. Four files are being added to the src/base/hmath
 folder.

The optimization is implemented so that it runs in a
separate thread called the impedanceModulationThread
in the motor controller.

review: https://reviewboard.rethink.cxm/r/8231/

resovles #18058
resolves #18506
resolves $18573