Fix squish issues during contact move in free space
This update fixes the issue of false squish getting triggered
when contact move nodes are used to move the robot with high
speed in free space. Because of the ambiguities in defining
squishes for contact moves where some amount of interaction
with environment is expected, we have decided to suppress
squishes during contact moves. So this update adds a new topic
for suppressing squish and interaction controller now publishes
to this topic only during contact moves in 100Hz. Additionally,
I have added fixes for some other bugs as well.

review: https://reviewboard.rethink.cxm/r/8458/

resolves #20142