Xenial/Kinetic Migration for Developer Workstations
Note: this is for Ubuntu 16.04 only, and not backwards compatible with
Ubuntu 14.04 workstations. Many thanks to Neil and Andrew for their help
and contributions throughout this process.
- minimum cmake version is now 3.0.2
- updated setupdev packages
- updated developer workstation apt resources
- updates nodejs to 8.9.3
- fixes the service_menu xenial build bugs
- bumps postgres to 9.5
- bumps OpenCV to 3.x
- general CMakeLists file cleanup in all Intera packages
- removed boost smart pointers in favor of C++11 std smart pointers
  where it was convenient
- updates build system from catkin_make to use the stricter, and
  parallel-core-compiling-capable catkin-tools

Tested building and launching Intera in sim on my workstation. Tested
setup-dev on two generic 16.04 workstations, then built Intera on them,
and launched in simulation. One stipulation is that one needs to
re-install sharp via NPM after the build, for now.

To test:

1) cd ~/intera
2) git checkout kinetic-xenial
3) ./bin/setup-dev
4) ./init -r sawyer_mp4 -s
5) catkin build
6) source devel/setup.sh
7) roscd js
8) npm install sharp
9) launch -s

No real robot-testing has been done as of yet.