Allow for arbitrary rotations for constrained zero-G
This update provides a new option for interaction control
which will allow for arbitrary rotations during constrained zero-G.

A boolean (rotations_for_constrained_zeroG) has been added to
interaction control command and this can be set True while the
commanded orientation is stationary. The actual orientation is
then allowed to change without any limits (only limited by the
joint position limits) with specified rotational stiffness.
For constrained zero-G applications where the stiffnesses for
translations and/or rotations along/about some axes are zero,
setting rotations_for_constrained_zeroG = True will provide
a greater flexibility with stabler behavior. Note that when
it is set to False, the range of rotations is limited to +/-
82 degree.