Fix bugs in Squish behavior during Contact
This update addresses a number of issues for squish behavior during contact.

Firstly, it fixes torque slewing bugs that led the arm to unstable behavior
after a squish. Now we reset torque slew counter when getting out of squish
only if arm is in interaction mode. Also, now we slew torquebias from the capped
torque when a squish is triggered, and we shortened the slewing duration from
250ms to 50ms.

Secondly, it ensures the engine to receive the collision state correctly when
a collision occurs. We now increased the publish rate from 20Hz to 100Hz, and
we latched the armIsStalled state so that the true state is guaranteed to be
published to the velocityControllerState topic. In addition, the engine now
no longer publishes redundant messages to interaction controller so that it
keeps a clean command history and brings a slight behavior improvement.

review: https://reviewboard.rethink.cxm/r/10013/
resolves IN-4551